move download file function to utilits
