+++ src/download_helper.js
@@ -1,5 +1,7 @@
 import Logger from "./logger";
 
+import { downloadFile } from "./utilities";
+
 const preamble = `#!/usr/bin/env bash
 # Generated by Bandcamp Enhancement Suite (https://github.com/sabjorn/BandcampEnhancementSuite)
 #
@@ -65,7 +67,7 @@ export default class DownloadHelper {
       const preamble = DownloadHelper.getDownloadPreamble();
       const downloadDocument = preamble + downloadList;
 
-      DownloadHelper.download(`bandcamp_${date}.txt`, downloadDocument);
+      downloadFile(`bandcamp_${date}.txt`, downloadDocument);
     });
   }
 
@@ -89,23 +91,6 @@ export default class DownloadHelper {
     filelist = filelist.substring(0, filelist.length - 6);
     filelist += "\n";
     return filelist;
-  }
-
-  static download(filename, text) {
-    var element = document.createElement("a");
-
-    element.setAttribute(
-      "href",
-      "data:text/plain;charset=utf-8," + encodeURIComponent(text)
-    );
-    element.setAttribute("download", filename);
-
-    element.style.display = "none";
-    document.body.appendChild(element);
-
-    element.click();
-
-    document.body.removeChild(element);
   }
 
   static dateString() {
