+++ src/utilities.js
@@ -105,3 +105,20 @@ export function getTralbumDetails(item_id, item_type = "a") {
 
   return fetch(`/api/mobile/25/tralbum_details`, requestOptions);
 }
+
+export function downloadFile(filename, text) {
+  var element = document.createElement("a");
+
+  element.setAttribute(
+    "href",
+    "data:text/plain;charset=utf-8," + encodeURIComponent(text)
+  );
+  element.setAttribute("download", filename);
+
+  element.style.display = "none";
+  document.body.appendChild(element);
+
+  element.click();
+
+  document.body.removeChild(element);
+}
